mv normalizationcache to own cpp/hpp files; mv rocksdbcache to own cpp/hpp files; update other files for appropriate includes"
