+++ src/rocksdbcache.hpp
@@ -1,0 +1,95 @@
+#ifndef __CARMEN_ROCKSDBCACHE_HPP__
+#define __CARMEN_ROCKSDBCACHE_HPP__
+
+#include "node_util.hpp"
+#include "cpp_util.hpp"
+#include <nan.h>
+
+namespace carmen {
+
+#define CACHE_MESSAGE 1
+#define CACHE_ITEM 1
+
+#define MEMO_PREFIX_LENGTH_T1 3
+#define MEMO_PREFIX_LENGTH_T2 6
+#define PREFIX_MAX_GRID_LENGTH 500000
+
+struct sortableGrid {
+    protozero::const_varint_iterator<uint64_t> it;
+    protozero::const_varint_iterator<uint64_t> end;
+    value_type unadjusted_lastval;
+    bool matches_language;
+};
+
+struct MergeBaton : carmen::noncopyable {
+    uv_work_t request;
+    std::string filename1;
+    std::string filename2;
+    std::string filename3;
+    std::string method;
+    std::string error;
+    Nan::Persistent<v8::Function> callback;
+};
+
+inline void decodeMessage(std::string const& message, intarray& array) {
+    protozero::pbf_reader item(message);
+    item.next(CACHE_ITEM);
+    auto vals = item.get_packed_uint64();
+    uint64_t lastval = 0;
+    // delta decode values.
+    for (auto it = vals.first; it != vals.second; ++it) {
+        if (lastval == 0) {
+            lastval = *it;
+            array.emplace_back(lastval);
+        } else {
+            lastval = lastval - *it;
+            array.emplace_back(lastval);
+        }
+    }
+}
+
+inline void decodeAndBoostMessage(std::string const& message, intarray& array) {
+    protozero::pbf_reader item(message);
+    item.next(CACHE_ITEM);
+    auto vals = item.get_packed_uint64();
+    uint64_t lastval = 0;
+    // delta decode values.
+    for (auto it = vals.first; it != vals.second; ++it) {
+        if (lastval == 0) {
+            lastval = *it;
+            array.emplace_back(lastval | LANGUAGE_MATCH_BOOST);
+        } else {
+            lastval = lastval - *it;
+            array.emplace_back(lastval | LANGUAGE_MATCH_BOOST);
+        }
+    }
+}
+
+class RocksDBCache : public node::ObjectWrap {
+  public:
+    ~RocksDBCache();
+    static Nan::Persistent<v8::FunctionTemplate> constructor;
+    static void Initialize(v8::Handle<v8::Object> target);
+    static NAN_METHOD(New);
+    static NAN_METHOD(pack);
+    static NAN_METHOD(merge);
+    static NAN_METHOD(list);
+    static NAN_METHOD(_get);
+    static NAN_METHOD(_getmatching);
+    static NAN_METHOD(_set);
+    static NAN_METHOD(coalesce);
+    explicit RocksDBCache();
+    void _ref() { Ref(); }
+    void _unref() { Unref(); }
+    std::shared_ptr<rocksdb::DB> db;
+};
+
+void mergeQueue(uv_work_t* req);
+void mergeAfter(uv_work_t* req);
+
+intarray __get(RocksDBCache const* c, std::string phrase, langfield_type langfield);
+intarray __getmatching(RocksDBCache const* c, std::string phrase, bool match_prefixes, langfield_type langfield);
+
+} // namespace carmen
+
+#endif // __CARMEN_ROCKSDBCACHE_HPP__
